Added mapping class, get read orientation does not work for merged SAM files
